+++ nncf/common/tensor_statistics/aggregator.py
@@ -31,6 +31,7 @@ from nncf.data.dataset import Dataset
 from nncf.data.dataset import ModelInput
 from nncf.experimental.common.tensor_statistics.statistics import TensorStatistic
 from nncf.tensor.tensor import TTensor
+from nncf.tensor.tensor import get_tensor_backend
 
 TensorType = TypeVar("TensorType")
 TModel = TypeVar("TModel")
@@ -103,7 +104,7 @@ class StatisticsAggregator(ABC, Generic[TTensor]):
 
         :param dir_path: The name of the directory from which to load the statistics.
         """
-        tensor_backend = statistics_serializer.get_tensor_backend(self.BACKEND)
+        tensor_backend = get_tensor_backend(self.BACKEND)
         loaded_data: Dict[str, Dict[str, TTensor]]
         metadata: Dict[str, Any]
         loaded_data, metadata = statistics_serializer.load_from_dir(dir_path, tensor_backend)
@@ -133,7 +134,7 @@ class StatisticsAggregator(ABC, Generic[TTensor]):
         """
         data_to_dump = self._prepare_statistics()
         metadata = {"backend": self.BACKEND.value, "subset_size": self.stat_subset_size}
-        tensor_backend = statistics_serializer.get_tensor_backend(self.BACKEND)
+        tensor_backend = get_tensor_backend(self.BACKEND)
         statistics_serializer.dump_to_dir(data_to_dump, dir_path, tensor_backend, metadata)
         nncf_logger.info(f"Statistics were successfully saved to a directory {dir_path}.")
