+++ nncf/tensor/tensor.py
@@ -11,8 +11,10 @@
 from __future__ import annotations
 
 import operator
-from typing import Any, Optional, Tuple, TypeVar, Union
+from typing import Any, Dict, Optional, Tuple, TypeVar, Union
 
+import nncf
+from nncf.common.utils.backend import BackendType
 from nncf.tensor.definitions import TensorBackend
 from nncf.tensor.definitions import TensorDataType
 from nncf.tensor.definitions import TensorDeviceType
@@ -251,3 +253,21 @@ def unwrap_tensor_data(obj: Any) -> TTensor:
     :return: The data of the Tensor object, or the object itself.
     """
     return obj.data if isinstance(obj, Tensor) else obj
+
+
+def get_tensor_backend(backend: BackendType) -> TensorBackend:
+    """
+    Returns a tensor backend based on the provided backend.
+    :param backend: Backend type.
+    :return: Corresponding tensor backend type.
+    """
+    BACKEND_TO_TENSOR_BACKEND: Dict[BackendType, TensorBackend] = {
+        BackendType.OPENVINO: TensorBackend.numpy,
+        BackendType.ONNX: TensorBackend.numpy,
+        BackendType.TORCH_FX: TensorBackend.torch,
+        BackendType.TORCH: TensorBackend.torch,
+    }
+    if backend not in BACKEND_TO_TENSOR_BACKEND:
+        raise nncf.ValidationError(f"Unsupported backend type: {backend}")
+
+    return BACKEND_TO_TENSOR_BACKEND[backend]
